Create the greeting component
